Add files via upload
This generates an image of the original Powerpuff Girls (link to source of image in code). The image is split into fourths and in each square, random shapes are generated. There is also user input, so when the user holds down the UP ARROW KEY and the DOWN ARROW KEY or drags the left mouse button, more shapes will be generated, the image will be generated but flipped, or the user can create the image with shapes.