Add files via upload
Want to show someone your appreciation or just want to brighten someone's day but have know idea how to? Well, here is an idea generator! There are three separate categories within the generator, so if you are thinking of wanting to go on a date but don't know how, press 'a.' If you want to give someone a gift but don't have the money, press 'b.' If you are okay with spending a few dollars on a gift, press'c.' An image will pop up and you can interpret it however you want. None of the images are mine.